requirements.txt
apturl==0.5.2
backports.weakref==1.0rc1
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
Brlapi==0.6.4
certifi==2017.7.27.1
chardet==3.0.4
checkbox-support==0.22
click==6.7
command-not-found==0.3
cookies==2.2.1
cryptography==1.2.3
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
dominate==2.3.1
feedparser==5.1.3
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Moment==0.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
googlemaps==2.5.1
guacamole==0.9.2
html5lib==0.9999999
httplib2==0.9.1
idna==2.6
imutils==0.4.3
itsdangerous==0.24
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==2.1.0
networkx==2.0
numpy==1.13.3
oauthlib==1.0.3
olefile==0.44
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==4.3.0
plainbox==0.25
protobuf==3.4.0
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.2.0
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-firebase==1.2
python-systemd==231
pytz==2017.3
PyWavelets==0.5.2
pyxdg==0.25
reportlab==3.3.0
requests==2.18.4
responses==0.8.1
scikit-image==0.13.1
scipy==1.0.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.11.0
SQLAlchemy==1.1.15
system-service==0.3
tensorflow-gpu==1.2.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.0
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
